refactor: SidebarMenuSection 및 SidebarTagSection에서 props 제거 및 내부 상태 관리로 변경
